+++ gtatool/src/cmds/dimension-add.cpp
@@ -33,6 +33,7 @@
 #include "opt.h"
 #include "cio.h"
 #include "str.h"
+#include "checked_cast.h"
 
 #include "lib.h"
 
@@ -130,7 +131,7 @@ extern "C" int gtatool_dimension_add(int argc, char *argv[])
                 // Write the GTA header
                 hdro.write_to(stdout);
                 // Manipulate the GTA data
-                blob element(hdri.element_size());
+                blob element(checked_cast<size_t>(hdri.element_size()));
                 gta::io_state si, so;
                 for (uintmax_t e = 0; e < hdri.elements(); e++)
                 {
